Miniter/Miniter_Web/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Twisted==0.1.2
PyJWT==1.4.0
Flask-Script